requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
openai==0.28.0
python-dotenv==1.0.1
